requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
joblib==1.0.1; python_version >= '3.6'
lightgbm==3.2.1
numpy==1.20.3
pandas==1.2.4
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2021.1
pyyaml==5.4.1
scikit-learn==0.24.2; python_version >= '3.6'
scipy==1.6.3
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sklearn==0.0
threadpoolctl==2.1.0; python_version >= '3.5'
wheel==0.36.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
